requirements-py27.txt
alembic==0.9.5
aniso8601==1.3.0
bcrypt==3.1.3
blinker==1.4
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.6
Flask==0.11.1
Flask-Admin==1.5.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.7
MarkupSafe==0.23
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
pycparser==2.18
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
uWSGI==2.0.14
Werkzeug==0.11.11
WTForms==2.1
